requirements.txt
requests>=2.28.0
pandas>=1.0.0
azure-storage-file-share>=12.0.0
